Added StudentID to tutor class and changed accountNumber to String.
StudentID was not apart of our Tutor model and it should be. I added a test for this as well. Account Number will be now stored as a string due to integer size and possible characters in the accountNumber. Also changed Tutor tests to use equals when checking strings. Finally changed some comments in the User class.